Move StringBinarySerializer tests to its own file
