+++ SAGESharpTests/IO/BinarySerializersReadingTests.cs
@@ -64,42 +64,6 @@ namespace SAGESharp.IO
         }
 
         [Test]
-        public void Test_Reading_A_String()
-        {
-            var expected = "hello world";
-
-            reader.ReadUInt32().Returns((uint)30);
-            reader.Position.Returns(50);
-            reader.ReadByte().Returns((byte)expected.Length);
-            reader.ReadBytes(expected.Length).Returns(new byte[]
-            {
-                // "hello"
-                0x68, 0x65, 0x6C, 0x6C, 0x6F,
-                // " "
-                0x20,
-                // "world"
-                0x77, 0x6F, 0x72, 0x6C, 0x64,
-            });
-
-            new StringBinarySerializer()
-                .Read(reader)
-                .Should()
-                .BeOfType<string>()
-                .Which
-                .Should()
-                .Be(expected);
-
-            Received.InOrder(() =>
-            {
-                reader.ReadUInt32();
-                reader.Position = 30;
-                reader.ReadByte();
-                reader.ReadBytes(expected.Length);
-                reader.Position = 50;
-            });
-        }
-
-        [Test]
         public void Test_Reading_A_List()
         {
             var expected = new List<char>() { 'a', 'b', 'c', 'd', 'e' };
