+++ SAGESharpTests/SAGESharpTests.csproj
@@ -90,6 +90,7 @@
     <Compile Include="IO\DefaultPropertyBinarySerializerFactoryTests.cs" />
     <Compile Include="IO\DefaultPropertyBinarySerializerTests.cs" />
     <Compile Include="IO\PrimitiveBinarySerializerTests.cs" />
+    <Compile Include="IO\StringBinarySerializerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SLB\ASCIIExtensionsTests.cs" />
     <Compile Include="SLB\Cinematic\Camera\CameraEqualityTests.cs" />
